requirements-dev.txt
appdirs==1.4.3
asgiref==3.2.3
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
djangorestframework==3.11.0
drf-nested-routers==0.91
filelock==3.0.12
importlib-metadata==1.5.0
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
virtualenv==20.0.4
zipp==2.2.0
